test(core): migrate UpdatePaneTests (Wave C)
52 tests, 175 failure sites preserved 1:1. The 19 compound `guard case` destructuring patterns convert to `Issue.record + return` to keep the failure-site count exact.
